requirements.txt
dash
dash-bootstrap-components
flask
openai
python-dotenv
gunicorn
